# packages/api-service/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Turnstile API",
    "description": "Token bridge system for Ethereum L1 and Aztec L2 networks",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {},
...
  },
  "paths": {
...
    "/contract/{address}": {
      "get": {
        "summary": "Get contract instance by address",
        "tags": [
          "Contracts"
        ],
        "description": "Get a contract instance by its address, optionally including artifact data",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "includeArtifact",
            "required": false,
            "description": "Include artifact data in the response (true/false)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "address",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "address": {
                      "type": "string",
                      "pattern": "^0x[a-fA-F0-9]{64}$"
                    },
                    "original_contract_class_id": {
                      "nullable": true,
                      "type": "string",
                      "pattern": "^0x[a-fA-F0-9]{64}$"
                    },
                    "current_contract_class_id": {
                      "nullable": true,
                      "type": "string",
                      "pattern": "^0x[a-fA-F0-9]{64}$"
                    },
                    "initialization_hash": {
                      "nullable": true,
                      "type": "string",
                      "pattern": "^0x[a-fA-F0-9]{64}$"
                    },
                    "deployment_params": {
                      "nullable": true,
                      "type": "object",
                      "properties": {
                        "constructorArtifact": {
                          "nullable": true,
                          "type": "string"
                        },
                        "constructorArgs": {
                          "nullable": true,
                          "type": "array",
                          "items": {}
                        },
                        "salt": {
                          "type": "string",
                          "pattern": "^0x[a-fA-F0-9]+$"
                        },
                        "publicKeys": {
                          "type": "string",
                          "pattern": "^0x[a-fA-F0-9]+$"
                        },
                        "deployer": {
                          "type": "string",
                          "pattern": "^0x[a-fA-F0-9]+$"
                        }
                      },
                      "required": [
                        "salt",
                        "publicKeys",
                        "deployer"
                      ],
                      "additionalProperties": false
                    },
                    "version": {
                      "type": "number"
                    },
                    "artifact_hash": {
                      "type": "string",
                      "pattern": "^0x[a-fA-F0-9]{64}$"
                    },
                    "artifact": {},
                    "contract_class_id": {
                      "type": "string",
                      "pattern": "^0x[a-fA-F0-9]{64}$"
                    }
                  },
                  "required": [
                    "address",
                    "original_contract_class_id",
                    "current_contract_class_id",
                    "initialization_hash",
                    "deployment_params",
                    "version"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/artifact/{identifier}": {
      "get": {
        "summary": "Get contract artifact by identifier",
        "tags": [
          "Contracts"
        ],
        "description": "Get a contract artifact by contract class ID or artifact hash",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "artifact_hash": {
                      "type": "string",
                      "pattern": "^0x[a-fA-F0-9]{64}$"
                    },
                    "artifact": {},
                    "contract_class_id": {
                      "type": "string",
                      "pattern": "^0x[a-fA-F0-9]{64}$"
                    }
                  },
                  "required": [
                    "artifact_hash",
                    "artifact",
                    "contract_class_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/contracts/by-class/{contractClassId}/addresses": {
      "get": {
        "summary": "Get contract addresses for a class",
        "tags": [
          "Contracts"
        ],
        "description": "Get all contract instance addresses that match the given contract class ID with configurable match scope",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "current",
                "original",
                "any"
              ]
            },
            "in": "query",
            "name": "match",
            "required": false,
            "description": "Match scope for contract instances: current (default), original, or any"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "contractClassId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "pattern": "^0x[a-fA-F0-9]{64}$"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Development server"
    }
  ],
...
}